+++ lib/widgets/custom_expansion_tile.dart
@@ -1,0 +1,50 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_platform_widgets/flutter_platform_widgets.dart';
+
+class CustomExpansionTile extends StatefulWidget {
+  final Widget title;
+  final List<Widget> children;
+
+  const CustomExpansionTile({
+    super.key,
+    required this.title,
+    required this.children,
+  });
+
+  @override
+  State<CustomExpansionTile> createState() => _CustomExpansionTileState();
+}
+
+class _CustomExpansionTileState extends State<CustomExpansionTile> {
+  bool _isExpanded = false;
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      mainAxisSize: MainAxisSize.min,
+      children: <Widget>[
+        PlatformListTile(
+          title: widget.title,
+          onTap: () {
+            setState(() {
+              _isExpanded = !_isExpanded;
+            });
+          },
+          trailing: Icon(
+            _isExpanded ? Icons.expand_less : Icons.expand_more,
+          ),
+        ),
+        AnimatedCrossFade(
+          firstChild: Container(),
+          secondChild: Column(
+            children: widget.children,
+          ),
+          crossFadeState: _isExpanded
+              ? CrossFadeState.showSecond
+              : CrossFadeState.showFirst,
+          duration: const Duration(milliseconds: 300),
+        ),
+      ],
+    );
+  }
+}
